feat: add request ID correlation for RPC calls
Added unique, monotonic request IDs to all JSON-RPC calls for improved
debugging and tracing. Request IDs now appear in logs and exception messages,
making it trivial to correlate errors with specific RPC calls.

Features:
- Extended RpcException to carry optional requestId field
- Request IDs appear as [requestId=X] prefix in exception messages
- All RPC logs include id=X parameter (e.g., [RPC] id=1 method=...)
- IDs are monotonically increasing per HttpBraneProvider instance
- Thread-safe ID generation via AtomicLong

Implementation:
- Updated HttpBraneProvider to extract numeric ID before string conversion
- Modified all RpcException constructors to accept requestId parameter
- Updated debug logging format to include request ID
- Fixed 4 ambiguous constructor overload issues with explicit (Long) null casts

Testing:
- Created RequestIdSanityCheck integration test
- Verifies request IDs in logs and exceptions
- Confirms monotonic ID sequence
- All unit tests (22/22) and integration tests pass

Example output:
  [RPC] id=1 method=eth_getBlockByNumber durationMicros=1234 ...
  [requestId=4] Network error during JSON-RPC call

This significantly improves debugging by making it easy to trace individual
requests through logs and correlate errors with specific RPC calls.

+++ brane-rpc/src/test/java/io/brane/rpc/DefaultPublicClientTest.java
@@ -118,7 +118,7 @@ class DefaultPublicClientTest {
                         resp.error().code(),
                         resp.error().message(),
                         resp.error().data() != null ? resp.error().data().toString() : null,
-                        null);
+                        (Long) null);
             }
             return resp;
         }
